+++ glue_vispy_viewers/tests/test_session_back_compat.py
@@ -4,11 +4,14 @@ from __future__ import absolute_import, division, print_function
 
 import os
 import pytest
+from mock import patch
 
 import numpy as np
 from numpy.testing import assert_equal
 
 from glue.core.state import GlueUnSerializer
+
+from ..volume.volume_viewer import QMessageBox
 
 DATA = os.path.join(os.path.dirname(__file__), 'data')
 
@@ -21,9 +24,10 @@ def test_scatter_volume(protocol):
     with open(filename, 'r') as f:
         session = f.read()
 
-    state = GlueUnSerializer.loads(session)
-
-    ga = state.object('__main__')
+    with patch.object(QMessageBox, 'question') as question:
+        question.return_value = QMessageBox.Yes
+        state = GlueUnSerializer.loads(session)
+        ga = state.object('__main__')
 
     dc = ga.session.data_collection
 
@@ -118,9 +122,10 @@ def test_scatter_volume_selection():
     with open(filename, 'r') as f:
         session = f.read()
 
-    state = GlueUnSerializer.loads(session)
-
-    ga = state.object('__main__')
+    with patch.object(QMessageBox, 'question') as question:
+        question.return_value = QMessageBox.Yes
+        state = GlueUnSerializer.loads(session)
+        ga = state.object('__main__')
 
     dc = ga.session.data_collection
 
@@ -150,5 +155,34 @@ def test_scatter_volume_selection():
 
     assert_equal(dc[0].subsets[0].to_mask(), expected_array)
     assert_equal(dc[1].subsets[0].to_mask(), expected_table)
+
+    ga.close()
+
+
+@pytest.mark.parametrize('protocol', [1, 2])
+def test_multiple_volumes(protocol):
+
+    # Before glue-vispy-viewers 0.12, volumes could be shown together without
+    # being linked, so when loading old session files we should suggest to
+    # auto link.
+
+    filename = os.path.join(DATA, 'multiple_volumes_v{0}.glu'.format(protocol))
+
+    with open(filename, 'r') as f:
+        session = f.read()
+
+    with patch.object(QMessageBox, 'question') as question:
+        question.return_value = QMessageBox.Yes
+        state = GlueUnSerializer.loads(session)
+        ga = state.object('__main__')
+
+    volume = ga.viewers[0][0]
+
+    assert volume.layers[0].enabled
+    assert volume.layers[1].enabled
+    assert volume.layers[2].enabled
+    assert volume.layers[3].enabled
+    assert volume.layers[4].enabled
+    assert volume.layers[5].enabled
 
     ga.close()