+++ CHANGES.md
@@ -3,7 +3,7 @@
 
 - Make it possible to view datasets that are linked but not on the same
   pixel grid together. Now also requires datasets to always be linked
-  in order to be shown in the same viewer. [#335]
+  in order to be shown in the same viewer. [#335, #337]
 
 0.11 (2018-11-14)
 -----------------