+++ CHANGES.md
@@ -7,6 +7,9 @@
 
 - Fix compatibility with the latest developer version of glue. [#339]
 
+- Fix bug with hidden layers in the 3D scatter viewer becoming visible after
+  saving and loading session file. [#340]
+
 0.11 (2018-11-14)
 -----------------